GenericTA
Library "GenericTA"

What is it?
The real generic library. Which means it is just covering most built-in indicators / functions, but with more parameters, so the user don't have to write more few lines to achieve something simple and replicative.

Development process?
Will tidy it up, and setting up in later stage.

Welcome to inbox me to improve the library ------
If you are finding a similar thing. That's a good news. Because I am making it.